Add a few more string methods
The methods padStart() and padEnd() are added. They add
the specified string as padding to the main string
at the front or the back up to a specified length.
The methods charAt() and charCodeAt() are also added. They provide
the capability to retrieve information on individual characters
making up the string. Property access [] on a string is also added,
which is similar to charAt().