Modify code to allow for multiple uses
The initial version allows for only a one-time use as the checkbox
is disabled after it is checked. The new version does not disable
the checkbox after it is checked, allowing for it to be reused
without having to refresh the web page. The message is updated
accordingly depending on whether the checkbox is checked or not.