Add example #4 of application of conditionals to building websites
This is a simple grade calculator which allows the user to input a score
and get their grade. It uses the switch statement to determine the range
in which the user's score fall and return the appropriate grade.
Moreover, the color of the text depends on whether the user's score results
in a passing (green) or failing (red) grade.