Add example #5 of application of conditionals to building websites
This is a simple season picker that allows the user to enter a month
and outputs the corresponding season. It uses the switch statement where
multiple cases (i.e. months) correspond to the same season.
It relies on the case statement continuing to execute code corresponding to
each case clause after the first matching clause until it reaches a break.