Add calendar generator
A calendar with the appropriate number of days is generated depending
on the month selected by the user.
The appearance of the calendar (i.e. HTML and CSS) is mostly finished.
The function createCalendar() that draws the calendar and displays the
correct month is created. An event handler to detect when the user
makes a selection is yet to be implemented.