Update the function updateTheme()
The <select> element is added as a parameter to the function for
greater reusability. Previously, the <select> element is hardcoded
as that for the dual theme case, thus it only works correctly for this
case. Now, it works for any case.
The function is also moved to the bottom of the script to improve code
organization.