Add multi-theme switcher to webpage-theme-switcher.html
The purpose of the two theme switchers is to show the applications of
a conditional operator and a switch statement, where the dual-theme switcher
uses the former while the multi-theme switcher uses the latter.
The box sizes are changed to ensure that the message remains on a single line
and some space is added between the boxes for better presentation.