Add exercise 3 from Odin JavaScript Fundamentals 2
It can be found at https://www.theodinproject.com/lessons/foundations-fundamentals-part-2.
The original exercise is on replit.com and outputs to the browser console,
but this version is modified to output to the browser screen.
Moreover, the original exercise runs on Node.js but this version runs
in the browser.