Explore JavaScript scoping
In particular, global scope vs. function scope is explored with variables.
Variables in the global scope are accessible from anywhere, including the console.
On the other hand, variables in a function scope is only available in the function
in which it is declared.
This rule extends beyond variables and also applies to functions, etc.