Add methods to add and remove elements from an array
The methods pop() and shift() work the same way, except that they work
on different sides of the array: pop() at the end and shift() at the start.
The same is true for the methods push() and unshift(): push() works at the end
of the array while unshift() at the start.