Update exercise 2 from 'Test your skills: Loops'
Previously, the name to be searched is hardcoded. The code is updated
to be more flexible by being able to receive a user-input name and then
outputting the name and the corresponding phone number.
In addition, when the user enters a name that is not in the phonebook,
a message stating that the name cannot be found is displayed.